Finished computing vertex normals

+++ CGlib/applications/trabalho2/1717553/Triangle.py
@@ -8,7 +8,7 @@ class Triangle(object):
         PQ = glm.vec3(vertices[0].x-vertices[1].x,vertices[0].y-vertices[1].y,vertices[0].z-vertices[1].z)
         RQ = glm.vec3(vertices[2].x-vertices[1].x,vertices[2].y-vertices[1].y,vertices[2].z-vertices[1].z)
         normal = glm.normalize(glm.cross(RQ,PQ))
-        self.normal = np.array([normal.x,normal.y,normal.z])
+        self.normal = np.array([normal.x,normal.y,normal.z], dtype=np.float32)
 
     def getCentroid(self):
         return np.array([self.vertices[0].x + self.vertices[1].x + self.vertices[2].x,
